+++ src/app/loadout-builder/process-worker/process.ts
@@ -1,3 +1,4 @@
+import { filterMap } from 'app/utils/collections';
 import { infoLog } from '../../utils/log';
 import {
   ArmorStatHashes,
@@ -19,9 +20,9 @@ import { SetTracker } from './set-tracker';
 import {
   AutoModData,
   LockedProcessMods,
-  ProcessArmorSet,
   ProcessItem,
   ProcessItemsByBucket,
+  ProcessResult,
   ProcessStatistics,
 } from './types';
 
@@ -46,14 +47,12 @@ export function process(
   /** Which artifice mods, large, and small stat mods are available */
   autoModOptions: AutoModData,
   /** Use stat mods to hit stat minimums */
-  autoStatMods: boolean
-): {
-  sets: ProcessArmorSet[];
-  combos: number;
-  /** The stat ranges of all sets that matched our filters & mod selection. */
-  statRangesFiltered?: StatRanges;
-  processInfo?: ProcessStatistics;
-} {
+  autoStatMods: boolean,
+  /** If set, only sets where at least one stat **exceeds** `resolvedStatConstraints` minimums will be returned */
+  strictUpgrades: boolean,
+  /** If set, LO will exit after finding at least one set that fits all constraints (and is a strict upgrade if `strictUpgrades` is set) */
+  stopOnFirstSet: boolean
+): ProcessResult {
   const pstart = performance.now();
 
   const statOrder = resolvedStatConstraints.map(({ statHash }) => statHash as ArmorStatHashes);
@@ -145,7 +144,7 @@ export function process(
     statistics: setStatistics,
   };
 
-  for (const helm of helms) {
+  itemLoop: for (const helm of helms) {
     for (const gaunt of gauntlets) {
       // For each additional piece, skip the whole branch if we've managed to get 2 exotics
       if (gaunt.isExotic && helm.isExotic) {
@@ -308,7 +307,7 @@ export function process(
 
             // We know this set satisfies all constraints.
             // Update highest possible reachable tiers.
-            updateMaxTiers(
+            const foundAnyImprovement = updateMaxTiers(
               precalculatedInfo,
               armor,
               stats,
@@ -392,6 +391,16 @@ export function process(
             // And now insert our set using the predicted tier. The rest of the stats string still uses the unboosted tiers but the error should be small
             tiersString = totalTier.toString(16) + tiersString;
             setTracker.insert(totalTier + predictedExtraTiers, tiersString, armor, stats);
+
+            if (stopOnFirstSet) {
+              if (strictUpgrades) {
+                if (foundAnyImprovement) {
+                  break itemLoop;
+                }
+              } else {
+                break itemLoop;
+              }
+            }
           }
         }
       }
@@ -400,7 +409,7 @@ export function process(
 
   const finalSets = setTracker.getArmorSets(RETURNED_ARMOR_SETS);
 
-  const sets = finalSets.map(({ armor, stats }) => {
+  const sets = filterMap(finalSets, ({ armor, stats }) => {
     // This only fails if minimum tier requirements cannot be hit, but we know they can because
     // we ensured it internally.
     const { mods, bonusStats } = pickOptimalStatMods(
@@ -413,12 +422,24 @@ export function process(
     const armorOnlyStats: Partial<ArmorStats> = {};
     const fullStats: Partial<ArmorStats> = {};
 
+    let hasStrictUpgrade = false;
+
     for (let i = 0; i < statOrder.length; i++) {
       const statHash = statOrder[i];
       const value = stats[i] + bonusStats[i];
       fullStats[statHash] = value;
 
+      if (strictUpgrades && !hasStrictUpgrade) {
+        const statFilter = resolvedStatConstraints[i];
+        const tier = Math.min(Math.max(Math.floor(value / 10), 0), 10);
+        hasStrictUpgrade ||= tier > statFilter.minTier;
+      }
+
       armorOnlyStats[statHash] = stats[i] - modStatsInStatOrder[i];
+    }
+
+    if (strictUpgrades && !hasStrictUpgrade) {
+      return undefined;
     }
 
     return {