+++ src/app/loadout-builder/process-worker/types.ts
@@ -1,4 +1,15 @@
-import { ArmorStatHashes, ArmorStats, LockableBucketHash } from '../types';
+import { ArmorStatHashes, ArmorStats, LockableBucketHash, StatRanges } from '../types';
+
+export interface ProcessResult {
+  /** A small number of the best sets, depending on operation mode. */
+  sets: ProcessArmorSet[];
+  /** The total number of combinations considered. */
+  combos: number;
+  /** The stat ranges of all sets that matched our filters & mod selection. */
+  statRangesFiltered?: StatRanges;
+  /** Statistics about how many sets passed/failed the constraints, for error reporting */
+  processInfo?: ProcessStatistics;
+}
 
 export interface ProcessItem {
   id: string;
@@ -46,7 +57,6 @@ export interface ProcessMod {
 /**
  * Data describing the mods that can be automatically picked.
  */
-
 export interface AutoModData {
   generalMods: {
     [key in ArmorStatHashes]?: {